improved phantomjs launcher

+++ tests/_bin/phantomjs_stop.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-kill `pgrep phantomjs` || true
+if pgrep "phantomjs" > /dev/null 2>&1
+then
+    kill `pgrep phantomjs` || true
+fi
